Move DummyLogger to it's own class so it can be used by other tests
